+++ updater/software/AcrobatReader2020.cs
@@ -1,6 +1,6 @@
 ﻿/*
     This file is part of the updater command line interface.
-    Copyright (C) 2022, 2023  Dirk Stolle
+    Copyright (C) 2022, 2023, 2024  Dirk Stolle
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -44,7 +44,7 @@ namespace updater.software
         /// <summary>
         /// expiration date for the publisher certificate
         /// </summary>
-        private static readonly DateTime certificateExpiration = new(2024, 5, 1, 23, 59, 59, DateTimeKind.Utc);
+        private static readonly DateTime certificateExpiration = new(2025, 11, 4, 23, 59, 59, DateTimeKind.Utc);
 
 
         /// <summary>
@@ -64,11 +64,11 @@ namespace updater.software
         /// details about the software.</returns>
         public override AvailableSoftware knownInfo()
         {
-            const string version = "20.005.30539";
+            const string version = "20.005.30574";
             var installer = new InstallInfoMsiPatch(
-                "https://ardownload2.adobe.com/pub/adobe/reader/win/Acrobat2020/2000530539/AcroRdr2020Upd2000530539_MUI.msp",
+                "https://ardownload2.adobe.com/pub/adobe/reader/win/Acrobat2020/2000530574/AcroRdr2020Upd2000530574_MUI.msp",
                 HashAlgorithm.SHA256,
-                "26d681ae6a171082712a79785116062d0bd284fe8034c4c8ae3d0a484efae246",
+                "c7e5e3a1620f1751b19113ae77c8250464c6b203f21be8358034e3a7dc8fe37a",
                 new Signature(publisherX509, certificateExpiration),
                 "/qn /norestart"
                 );